change rule priorities to run simplifiers first; add reify
Create a rule priority taxonomy, and refactor rules to follow this new
convention. In particular, ensure that partial and total evaluation and
simplification rules happen before representation changes and solver
specific rules.

DESCRIPTION

Simplifying a model before applying representation changes should ensure
that nested sums do not get propagated across the model. For example,
the transformation for min turns it into many binary constraints. If
simplification was not done first, simplification steps would have to be
ran on each of these binary constraints, instead of just on the input
model.

Ensuring that everything is always in canonical forms should stop rules
looping, and help with pattern matching. The implementation of
canonical forms is incomplete and outside the scope of this patch.

These rule changes caused basic-div-04 to fail. To fix this, this patch
also does the following:

- add reify constraint (reify(X,1) if constraint X is SAT, reify(X,0) if
  constraint X is UNSAT)

- add rule not(X) ~> reify(X,0) (where X is a constraint), allowing
  negated constraints to be passed to Minion.

MOTIVATION

The commit c2b74602c (add broken complex partial evaluator test
(arithmetic, nesting, min), 2024-10-12) added a test case for the
partial evaluator that never terminated.

Simplification steps should be ran before the decomposition of the min
constraint and the conversion of expressions into more specialised /
solver specific forms (such as ineq, sumgeq, etc); however, this was not
the case in this test. In this test, ineq / sumgeq were being applied
before the sums had been flattened.

After this refactor, this test now terminates with failure. A flattening
rule for ineq is needed before this test passes and will be added
in a later commit.

+++ crates/conjure_core/src/rule_engine/rule_set.rs
@@ -33,9 +33,9 @@ pub struct RuleSet<'a> {
     pub name: &'a str,
     /// Order of the RuleSet. Used to establish a consistent order of operations when resolving rules.
     /// If two RuleSets overlap (contain the same rule but with different priorities), the RuleSet with the higher order will be used as the source of truth.
-    pub order: u8,
+    pub order: u16,
     /// A map of rules to their priorities. This will be lazily initialized at runtime.
-    rules: OnceLock<HashMap<&'a Rule<'a>, u8>>,
+    rules: OnceLock<HashMap<&'a Rule<'a>, u16>>,
     /// The names of the rule sets that this rule set depends on.
     dependency_rs_names: &'a [&'a str],
     dependencies: OnceLock<HashSet<&'a RuleSet<'a>>>,
@@ -46,7 +46,7 @@ pub struct RuleSet<'a> {
 impl<'a> RuleSet<'a> {
     pub const fn new(
         name: &'a str,
-        order: u8,
+        order: u16,
         dependencies: &'a [&'a str],
         solver_families: &'a [SolverFamily],
     ) -> Self {
@@ -61,8 +61,8 @@ impl<'a> RuleSet<'a> {
     }
 
     /// Get the rules of this rule set, evaluating them lazily if necessary
-    /// Returns a `&HashMap<&Rule, u8>` where the key is the rule and the value is the priority of the rule.
-    pub fn get_rules(&self) -> &HashMap<&'a Rule<'a>, u8> {
+    /// Returns a `&HashMap<&Rule, u16>` where the key is the rule and the value is the priority of the rule.
+    pub fn get_rules(&self) -> &HashMap<&'a Rule<'a>, u16> {
         match self.rules.get() {
             None => {
                 let rules = self.resolve_rules();
@@ -104,12 +104,12 @@ impl<'a> RuleSet<'a> {
     }
 
     /// Resolve the rules of this rule set ("reverse the arrows")
-    fn resolve_rules(&self) -> HashMap<&'a Rule<'a>, u8> {
+    fn resolve_rules(&self) -> HashMap<&'a Rule<'a>, u16> {
         let mut rules = HashMap::new();
 
         for rule in get_rules() {
             let mut found = false;
-            let mut priority: u8 = 0;
+            let mut priority: u16 = 0;
 
             for (name, p) in rule.rule_sets {
                 if *name == self.name {
